Major restructure.
Bug: Properties of EntityItem aren't being persisted.

The aim is to reduce complexity and inheritance by moving compelxity
out into new CharacterSheet and PlayableState classes.

Creation of a CharacterSheet class clearly separates where a player's
charactersheet starts and ends.

Creation of the PlayableState class clearly separates the run time
state of a Player or NPC in the game.

Deleted Race class.

Spellcast class renamed to Casting.