Reworked CombatArena to be the source of valid combat actions.
In the fifthed package ...

Remove the chain-style initialisation for CombatArena.
Have the CombatArena hold the PlayerCharacter performing the actions.
Remove the link from the CharacterClass sub-types to the PlayerCharacter.
Remove the link from the Spellcaster interface to the PlayerCharacter.
Remove the chain-style initialisation for CharacterClass.
Remove name from CharacterClass and sub-types as the type is the name.
Simplify the constructor for the CharacterClass sub-types Fighter and Magus.
Casting class now knows the caster which is used in validation.
Restructured unit tests.
Added TestHelper to created useful PlayerCharacter objects.